wip: Are the Binance URLs the same startum server?
I've modified this to try to find out if the two Biance Pool URLs
point to the same startum server. This is relevant as Binance
sometimes includes P2TR spends and sometimes not. This might be a way
to find out.

However, I couldn't really tell as they sometimes send the same work
and sometimes do not. I decided not to spend any more time on it for
now.